refactor(backend): carve out a real domain layer
Move the I/O-free logic out of core/services and core/audio into backend/domain:
titles, formatting (was list_format), filenames (was filename_parser) and the
suggestion engine with its suggesters.

Split core/audio/tags.py along its existing seam: the tag registry, StarlibMeta
and TrackInfo are pure and land in domain/tags.py; TrackHandler is a mutagen +
ffmpeg adapter and moves to infra/audio/track_handler.py alongside folders.py.

Break both suggestion cycles by giving the suggester protocol and context their
own module (domain/suggestions/types.py) and moving REGISTRY to registry.py, so
the dependency runs engine -> registry -> suggesters -> types in one direction
and the three function-local imports are gone.

Delete the empty core/domain package.
