+++ backend/api/metadata/collection.py
@@ -7,8 +7,8 @@ from typing import Annotated
 from fastapi import APIRouter, Depends, HTTPException, status
 
 from backend.api.deps import get_root_folder
-from backend.core.audio.folders import FolderHandler
 from backend.core.services import collection
+from backend.infra.audio.folders import FolderHandler
 from backend.schemas.metadata import CollectionSoundcloudIdsResponse, CollectionStatsResponse
 
 logger = logging.getLogger(__name__)
