--- backend/core/audio/__init__.py
@@ -1,1 +1,0 @@
-"""Local audio-file metadata: tag I/O, folder discovery, title heuristics."""
