--- backend/core/domain/__init__.py
@@ -1,1 +1,0 @@
-"""Domain logic - pure functions."""
