+++ backend/domain/suggestions/helpers.py
