+++ backend/domain/suggestions/suggesters/artist.py
@@ -15,10 +15,10 @@ have a SC track with one artist and a filename with another.
 
 from __future__ import annotations
 
-from backend.core.audio.titles import rank_artists
-from backend.core.services.list_format import aggregate_distinct, normalize_list, split_list
-from backend.core.services.suggesters._base import candidate, compact
-from backend.core.services.suggestion_engine import SuggestionContext
+from backend.domain.formatting import aggregate_distinct, normalize_list, split_list
+from backend.domain.suggestions.helpers import candidate, compact
+from backend.domain.suggestions.types import SuggestionContext
+from backend.domain.titles import rank_artists
 from backend.schemas.suggestions import FieldName, FieldSuggestion, SuggestionSource
