+++ backend/domain/suggestions/suggesters/bpm_key.py
@@ -7,8 +7,8 @@ one-click accept. BPM is rounded to int to match the local tag column.
 
 from __future__ import annotations
 
-from backend.core.services.suggesters._base import candidate, compact
-from backend.core.services.suggestion_engine import SuggestionContext
+from backend.domain.suggestions.helpers import candidate, compact
+from backend.domain.suggestions.types import SuggestionContext
 from backend.schemas.suggestions import FieldName, FieldSuggestion
