+++ backend/domain/suggestions/suggesters/genre.py
@@ -7,8 +7,8 @@ forget the dedicated genre slot).
 
 from __future__ import annotations
 
-from backend.core.services.suggesters._base import candidate, compact
-from backend.core.services.suggestion_engine import SuggestionContext
+from backend.domain.suggestions.helpers import candidate, compact
+from backend.domain.suggestions.types import SuggestionContext
 from backend.schemas.suggestions import FieldName, FieldSuggestion
