+++ backend/domain/suggestions/suggesters/mix_name.py
@@ -8,8 +8,8 @@ is only shown when the field is empty or different.
 
 from __future__ import annotations
 
-from backend.core.services.suggesters._base import candidate, compact
-from backend.core.services.suggestion_engine import SuggestionContext
+from backend.domain.suggestions.helpers import candidate, compact
+from backend.domain.suggestions.types import SuggestionContext
 from backend.schemas.suggestions import FieldName, FieldSuggestion
