+++ backend/domain/suggestions/suggesters/release.py
@@ -14,8 +14,8 @@ from __future__ import annotations
 
 from datetime import date, datetime
 
-from backend.core.services.suggesters._base import candidate, compact
-from backend.core.services.suggestion_engine import SuggestionContext
+from backend.domain.suggestions.helpers import candidate, compact
+from backend.domain.suggestions.types import SuggestionContext
 from backend.schemas.suggestions import FieldName, FieldSuggestion, SCTrackPayload
